Integration tests closes #21

+++ SocialNetworkApi.Logic/Implementations/UsersService.cs
@@ -131,7 +131,10 @@ namespace SocialNetworkApi.Services.Implementations
 
             await _userManager.UpdateAsync(user);
             await _userManager.SetLockoutEnabledAsync(user, true);
-            await _profilesService.DeleteByIdAsync(user.Profile.Id.ToString());
+            if(user.Profile != null)
+            {
+                await _profilesService.DeleteByIdAsync(user.Profile.Id.ToString());
+            }
 
             return true;
         }
@@ -156,7 +159,10 @@ namespace SocialNetworkApi.Services.Implementations
             user.IsDeleted = false;
             await _userManager.UpdateAsync(user);
             await _userManager.SetLockoutEnabledAsync(user, false);
-            await _profilesService.ReinstateAsync(user.Profile.Id.ToString());
+            if(user.Profile != null)
+            {
+                await _profilesService.ReinstateAsync(user.Profile.Id.ToString());
+            }
 
             return true;
         }
